Move ModelFromDeploymentEndpoint class to public
